// client/src/styles/theme.ts
// const colors = {
//   primaryBlue: "#2A788D",
//   black: "#000000",
//   lightBlueGray: "#8CB6CF",
//   darkGreenish: "#426F5F",
//   goldYellow: "#E4C75D",
//   graySilver: "#B5B8B9",
//   darkGray: "#7C7E80",
// };
export const colors = {
  Primary: "#2A788D",       // טורקיז כהה – כפתורים וכותרות
  AccentLight: "#8CB6CF",   // כחול-אפרפר – טקסטים דקורטיביים
  TextGrPrimaryay: "#7C7E80",      // אפור כהה – טקסט לייבלים
  Error: "#E4C75D",         // זהב – שגיאות (אפשר גם אדום אם תעדיפי)
  Link: "#426F5F",          // ירקרק – קישורים
  Black: "#000000",         // שחור
  White: "#ffffff",         // רקע
};


export const fonts = {
  base: `"Varela Round", sans-serif`,
  heading: `"Rubik", sans-serif`,
};